+++ src/app/dashboards/scripted.js
@@ -22,9 +22,6 @@ var dashboard, timspan;
 // All url parameters are available via the ARGS object
 var ARGS;
 
-// Set a default timespan if one isn't specified
-timspan = '1d';
-
 // Intialize a skeleton with nothing but a rows array and service object
 dashboard = {
   rows : [],
@@ -32,8 +29,12 @@ dashboard = {
 
 // Set a title
 dashboard.title = 'Scripted dash';
+
+// set default time
+// time can be overriden in the url using from/to parameteres, but this is
+// handled automatically in grafana core during dashboard initialization
 dashboard.time = {
-  from: "now-" + (ARGS.from || timspan),
+  from: 'now-6h',
   to: "now"
 };