+++ src/app/directives/panelMenu.js
@@ -72,7 +72,7 @@ function (angular, $, _) {
             $link.toggleClass('has-panel-links', showIcon);
           });
 
-          function dismiss(time) {
+          function dismiss(time, force) {
             clearTimeout(timeout);
             timeout = null;
 
@@ -82,9 +82,11 @@ function (angular, $, _) {
             }
 
             // if hovering or draging pospone close
-            if ($menu.is(':hover') || $scope.dashboard.$$panelDragging) {
-              dismiss(2500);
-              return;
+            if (force !== true) {
+              if ($menu.is(':hover') || $scope.dashboard.$$panelDragging) {
+                dismiss(2200);
+                return;
+              }
             }
 
             if (menuScope) {
@@ -97,7 +99,12 @@ function (angular, $, _) {
             }
           }
 
-          var showMenu = function() {
+          var showMenu = function(e) {
+            // if menu item is clicked and menu was just removed from dom ignore this event
+            if (!$.contains(document, e.target)) {
+              return;
+            }
+
             if ($menu) {
               dismiss();
               return;
@@ -124,6 +131,9 @@ function (angular, $, _) {
 
             menuScope = $scope.$new();
             menuScope.extendedMenu = getExtendedMenu($scope);
+            menuScope.dismiss = function() {
+              dismiss(null, true);
+            };
 
             $('.panel-menu').remove();
             elem.append($menu);
@@ -134,7 +144,7 @@ function (angular, $, _) {
             $(".panel-container").removeClass('panel-highlight');
             $panelContainer.toggleClass('panel-highlight');
 
-            dismiss(2500);
+            dismiss(2200);
           };
 
           if ($scope.panelMeta.titlePos && $scope.panel.title) {